+++ src/cargo/util/mod.rs
@@ -9,6 +9,7 @@ pub use self::hex::{to_hex, short_hash};
 pub use self::pool::TaskPool;
 pub use self::dependency_queue::{DependencyQueue, Fresh, Dirty, Freshness};
 pub use self::graph::Graph;
+pub use self::to_url::ToUrl;
 
 pub mod graph;
 pub mod process_builder;
@@ -21,3 +22,4 @@ pub mod errors;
 pub mod hex;
 mod pool;
 mod dependency_queue;
+mod to_url;
